Restructure base Change classes
