+++ packages/apollo-collaboration-server/src/changes/changes.service.ts
@@ -77,12 +77,10 @@ export class ChangesService {
         throw new UnprocessableEntityException(String(e))
       }
 
-      // Add change information to change -collection
-      this.logger.debug(`ChangeIds: ${change.changedIds}`)
-      this.logger.debug(`AssemblyId: ${change.assemblyId}`)
-
       // Add entry to change collection
       const changeEntry: CreateChangeDto = {
+        // eslint-disable-next-line @typescript-eslint/ban-ts-comment
+        // @ts-ignore
         assembly: change.assemblyId,
         ...change,
         user: 'demo user id',
