+++ packages/apollo-collaboration-server/src/changes/dto/create-change.dto.ts
@@ -1,6 +1,6 @@
 export class CreateChangeDto {
-  readonly assembly: string
-  readonly changedIds: string[]
+  readonly assembly?: string
+  readonly changedIds?: string[]
   readonly reverts?: string
   readonly user: string
   readonly typeName: string
