+++ packages/apollo-shared/src/BackendDrivers/BackendDriver.ts
@@ -1,8 +1,11 @@
 import { Region } from '@jbrowse/core/util'
 import { AnnotationFeatureSnapshot } from 'apollo-mst'
 
-import { SubmitOpts } from '../ChangeManager'
-import { Change, ClientDataStore } from '../ChangeManager/Change'
+import { SubmitOpts } from '../ChangeManager/ChangeManager'
+import {
+  Change,
+  ClientDataStore,
+} from '../ChangeManager/changes/abstract/Change'
 import { ValidationResultSet } from '../Validations/ValidationSet'
 
 export abstract class BackendDriver {
