+++ packages/apollo-shared/src/ChangeManager/changes/LocationEndChange.ts
@@ -5,47 +5,50 @@ import {
   ChangeOptions,
   ClientDataStore,
   LocalGFF3DataStore,
-  SerializedChange,
   ServerDataStore,
-} from './Change'
-import { FeatureChange } from './FeatureChange'
+} from './abstract/Change'
+import {
+  FeatureChange,
+  SerializedFeatureChange,
+} from './abstract/FeatureChange'
 
-interface SerializedTypeChangeBase extends SerializedChange {
-  typeName: 'TypeChange'
+interface SerializedLocationEndChangeBase extends SerializedFeatureChange {
+  typeName: 'LocationEndChange'
 }
 
-interface TypeChangeDetails {
+export interface LocationEndChangeDetails {
   featureId: string
-  oldType: string
-  newType: string
+  oldEnd: number
+  newEnd: number
 }
 
-interface SerializedTypeChangeSingle
-  extends SerializedTypeChangeBase,
-    TypeChangeDetails {}
+interface SerializedLocationEndChangeSingle
+  extends SerializedLocationEndChangeBase,
+    LocationEndChangeDetails {}
 
-interface SerializedTypeChangeMultiple extends SerializedTypeChangeBase {
-  changes: TypeChangeDetails[]
+interface SerializedLocationEndChangeMultiple
+  extends SerializedLocationEndChangeBase {
+  changes: LocationEndChangeDetails[]
 }
 
-type SerializedTypeChange =
-  | SerializedTypeChangeSingle
-  | SerializedTypeChangeMultiple
+type SerializedLocationEndChange =
+  | SerializedLocationEndChangeSingle
+  | SerializedLocationEndChangeMultiple
 
-export class TypeChange extends FeatureChange {
-  typeName = 'TypeChange' as const
-  changes: TypeChangeDetails[]
+export class LocationEndChange extends FeatureChange {
+  typeName = 'LocationEndChange' as const
+  changes: LocationEndChangeDetails[]
 
-  constructor(json: SerializedTypeChange, options?: ChangeOptions) {
+  constructor(json: SerializedLocationEndChange, options?: ChangeOptions) {
     super(json, options)
     this.changes = 'changes' in json ? json.changes : [json]
   }
 
-  toJSON(): SerializedTypeChange {
+  toJSON(): SerializedLocationEndChange {
     const { changes, changedIds, typeName, assemblyId } = this
     if (changes.length === 1) {
-      const [{ featureId, oldType, newType }] = changes
-      return { typeName, changedIds, assemblyId, featureId, oldType, newType }
+      const [{ featureId, oldEnd, newEnd }] = changes
+      return { typeName, changedIds, assemblyId, featureId, oldEnd, newEnd }
     }
     return { typeName, changedIds, assemblyId, changes }
   }
@@ -63,8 +66,8 @@ export class TypeChange extends FeatureChange {
       topLevelFeature: FeatureDocument
     }[] = []
     // Let's first check that all features are found and those old values match with expected ones. We do this just to be sure that all changes can be done.
-    for (const entry of changes) {
-      const { featureId, oldType } = entry
+    for (const change of changes) {
+      const { featureId, oldEnd } = change
 
       // Search correct feature
       const topLevelFeature = await featureModel
@@ -87,8 +90,8 @@ export class TypeChange extends FeatureChange {
         throw new Error(errMsg)
       }
       logger.debug?.(`*** Found feature: ${JSON.stringify(foundFeature)}`)
-      if (foundFeature.type !== oldType) {
-        const errMsg = `*** ERROR: Feature's current type "${topLevelFeature.type}" doesn't match with expected value "${oldType}"`
+      if (foundFeature.end !== oldEnd) {
+        const errMsg = `*** ERROR: Feature's current end value ${foundFeature.end} doesn't match with expected value ${oldEnd}`
         logger.error(errMsg)
         throw new Error(errMsg)
       }
@@ -97,11 +100,11 @@ export class TypeChange extends FeatureChange {
 
     // Let's update objects.
     for (const [idx, change] of changes.entries()) {
-      const { newType } = change
+      const { newEnd } = change
       const { feature, topLevelFeature } = featuresForChanges[idx]
-      feature.type = newType
+      feature.end = newEnd
       if (topLevelFeature._id.equals(feature._id)) {
-        topLevelFeature.markModified('type') // Mark as modified. Without this save() -method is not updating data in database
+        topLevelFeature.markModified('end') // Mark as modified. Without this save() -method is not updating data in database
       } else {
         topLevelFeature.markModified('children') // Mark as modified. Without this save() -method is not updating data in database
       }
@@ -133,7 +136,7 @@ export class TypeChange extends FeatureChange {
       if (!feature) {
         throw new Error(`Could not find feature with identifier "${changedId}"`)
       }
-      feature.setType(this.changes[idx].newType)
+      feature.setEnd(this.changes[idx].newEnd)
     })
   }
 
@@ -145,10 +148,10 @@ export class TypeChange extends FeatureChange {
       .reverse()
       .map((endChange) => ({
         featureId: endChange.featureId,
-        oldType: endChange.newType,
-        newType: endChange.oldType,
+        oldEnd: endChange.newEnd,
+        newEnd: endChange.oldEnd,
       }))
-    return new TypeChange(
+    return new LocationEndChange(
       {
         changedIds: inverseChangedIds,
         typeName,
@@ -158,4 +161,10 @@ export class TypeChange extends FeatureChange {
       { logger },
     )
   }
+}
+
+export function isLocationEndChange(
+  change: unknown,
+): change is LocationEndChange {
+  return (change as LocationEndChange).typeName === 'LocationEndChange'
 }
