+++ packages/apollo-shared/src/ChangeManager/changes/TypeChange.ts
@@ -5,48 +5,49 @@ import {
   ChangeOptions,
   ClientDataStore,
   LocalGFF3DataStore,
-  SerializedChange,
   ServerDataStore,
-} from './Change'
-import { FeatureChange } from './FeatureChange'
+} from './abstract/Change'
+import {
+  FeatureChange,
+  SerializedFeatureChange,
+} from './abstract/FeatureChange'
 
-interface SerializedLocationEndChangeBase extends SerializedChange {
-  typeName: 'LocationEndChange'
+interface SerializedTypeChangeBase extends SerializedFeatureChange {
+  typeName: 'TypeChange'
 }
 
-export interface LocationEndChangeDetails {
+interface TypeChangeDetails {
   featureId: string
-  oldEnd: number
-  newEnd: number
+  oldType: string
+  newType: string
 }
 
-interface SerializedLocationEndChangeSingle
-  extends SerializedLocationEndChangeBase,
-    LocationEndChangeDetails {}
+interface SerializedTypeChangeSingle
+  extends SerializedTypeChangeBase,
+    TypeChangeDetails {}
 
-interface SerializedLocationEndChangeMultiple
-  extends SerializedLocationEndChangeBase {
-  changes: LocationEndChangeDetails[]
+interface SerializedTypeChangeMultiple extends SerializedTypeChangeBase {
+  changes: TypeChangeDetails[]
 }
 
-type SerializedLocationEndChange =
-  | SerializedLocationEndChangeSingle
-  | SerializedLocationEndChangeMultiple
+type SerializedTypeChange =
+  | SerializedTypeChangeSingle
+  | SerializedTypeChangeMultiple
 
-export class LocationEndChange extends FeatureChange {
-  typeName = 'LocationEndChange' as const
-  changes: LocationEndChangeDetails[]
+export class TypeChange extends FeatureChange {
+  typeName = 'TypeChange' as const
+  changes: TypeChangeDetails[]
 
-  constructor(json: SerializedLocationEndChange, options?: ChangeOptions) {
+  constructor(json: SerializedTypeChange, options?: ChangeOptions) {
     super(json, options)
     this.changes = 'changes' in json ? json.changes : [json]
   }
 
-  toJSON(): SerializedLocationEndChange {
+  toJSON(): SerializedTypeChange {
     const { changes, changedIds, typeName, assemblyId } = this
     if (changes.length === 1) {
-      const [{ featureId, oldEnd, newEnd }] = changes
-      return { typeName, changedIds, assemblyId, featureId, oldEnd, newEnd }
+      const [{ featureId, oldType, newType }] = changes
+      return { typeName, changedIds, assemblyId, featureId, oldType, newType }
     }
     return { typeName, changedIds, assemblyId, changes }
   }
@@ -64,8 +65,8 @@ export class LocationEndChange extends FeatureChange {
       topLevelFeature: FeatureDocument
     }[] = []
     // Let's first check that all features are found and those old values match with expected ones. We do this just to be sure that all changes can be done.
-    for (const change of changes) {
-      const { featureId, oldEnd } = change
+    for (const entry of changes) {
+      const { featureId, oldType } = entry
 
       // Search correct feature
       const topLevelFeature = await featureModel
@@ -88,8 +89,8 @@ export class LocationEndChange extends FeatureChange {
         throw new Error(errMsg)
       }
       logger.debug?.(`*** Found feature: ${JSON.stringify(foundFeature)}`)
-      if (foundFeature.end !== oldEnd) {
-        const errMsg = `*** ERROR: Feature's current end value ${foundFeature.end} doesn't match with expected value ${oldEnd}`
+      if (foundFeature.type !== oldType) {
+        const errMsg = `*** ERROR: Feature's current type "${topLevelFeature.type}" doesn't match with expected value "${oldType}"`
         logger.error(errMsg)
         throw new Error(errMsg)
       }
@@ -98,11 +99,11 @@ export class LocationEndChange extends FeatureChange {
 
     // Let's update objects.
     for (const [idx, change] of changes.entries()) {
-      const { newEnd } = change
+      const { newType } = change
       const { feature, topLevelFeature } = featuresForChanges[idx]
-      feature.end = newEnd
+      feature.type = newType
       if (topLevelFeature._id.equals(feature._id)) {
-        topLevelFeature.markModified('end') // Mark as modified. Without this save() -method is not updating data in database
+        topLevelFeature.markModified('type') // Mark as modified. Without this save() -method is not updating data in database
       } else {
         topLevelFeature.markModified('children') // Mark as modified. Without this save() -method is not updating data in database
       }
@@ -134,7 +135,7 @@ export class LocationEndChange extends FeatureChange {
       if (!feature) {
         throw new Error(`Could not find feature with identifier "${changedId}"`)
       }
-      feature.setEnd(this.changes[idx].newEnd)
+      feature.setType(this.changes[idx].newType)
     })
   }
 
@@ -146,10 +147,10 @@ export class LocationEndChange extends FeatureChange {
       .reverse()
       .map((endChange) => ({
         featureId: endChange.featureId,
-        oldEnd: endChange.newEnd,
-        newEnd: endChange.oldEnd,
+        oldType: endChange.newType,
+        newType: endChange.oldType,
       }))
-    return new LocationEndChange(
+    return new TypeChange(
       {
         changedIds: inverseChangedIds,
         typeName,
@@ -159,10 +160,4 @@ export class LocationEndChange extends FeatureChange {
       { logger },
     )
   }
-}
-
-export function isLocationEndChange(
-  change: unknown,
-): change is LocationEndChange {
-  return (change as LocationEndChange).typeName === 'LocationEndChange'
 }
