+++ packages/apollo-shared/src/ChangeManager/changes/abstract/AssemblySpecificChange.ts
@@ -1,40 +1,25 @@
 import { GFF3Feature } from '@gmod/gff'
 import type { AnnotationFeatureSnapshot } from 'apollo-mst'
-import { Feature, FileDocument, RefSeqDocument } from 'apollo-schemas'
+import { FileDocument, RefSeqDocument } from 'apollo-schemas'
 import ObjectID from 'bson-objectid'
-import type { Types } from 'mongoose'
 
-import { Change, ServerDataStore } from './Change'
+import {
+  Change,
+  ChangeOptions,
+  SerializedChange,
+  ServerDataStore,
+} from './Change'
 
-export abstract class FeatureChange extends Change {
-  /**
-   * Get single feature by featureId
-   * @param feature -
-   * @param featureId -
-   * @returns
-   */
-  getFeatureFromId(feature: Feature, featureId: string): Feature | null {
-    const { logger } = this
-    logger.verbose?.(`Entry=${JSON.stringify(feature)}`)
+export interface SerializedAssemblySpecificChange extends SerializedChange {
+  assemblyId: string
+}
 
-    if (feature._id.equals(featureId)) {
-      logger.debug?.(
-        `Top level featureId matches in the object ${JSON.stringify(feature)}`,
-      )
-      return feature
-    }
-    // Check if there is also childFeatures in parent feature and it's not empty
-    // Let's get featureId from recursive method
-    logger.debug?.(
-      `FeatureId was not found on top level so lets make recursive call...`,
-    )
-    for (const [, childFeature] of feature.children || new Map()) {
-      const subFeature = this.getFeatureFromId(childFeature, featureId)
-      if (subFeature) {
-        return subFeature
-      }
-    }
-    return null
+export abstract class AssemblySpecificChange extends Change {
+  assemblyId: string
+
+  constructor(json: SerializedAssemblySpecificChange, options?: ChangeOptions) {
+    super(json, options)
+    this.assemblyId = json.assemblyId
   }
 
   async addRefSeqIntoDb(
@@ -192,7 +177,7 @@ export abstract class FeatureChange extends Change {
 
   async addFeatureIntoDb(gff3Feature: GFF3Feature, backend: ServerDataStore) {
     const { featureModel, refSeqModel, session } = backend
-    const { assemblyId, logger } = this
+    const { assemblyId, logger, refSeqCache } = this
 
     for (const featureLine of gff3Feature) {
       const { seq_id: refName } = featureLine
@@ -201,7 +186,7 @@ export abstract class FeatureChange extends Change {
           `Valid seq_id not found in feature ${JSON.stringify(featureLine)}`,
         )
       }
-      let refSeqDoc = this.refSeqCache.get(refName)
+      let refSeqDoc = refSeqCache.get(refName)
       if (!refSeqDoc) {
         refSeqDoc =
           (await refSeqModel
@@ -209,7 +194,7 @@ export abstract class FeatureChange extends Change {
             .session(session)
             .exec()) || undefined
         if (refSeqDoc) {
-          this.refSeqCache.set(refName, refSeqDoc)
+          refSeqCache.set(refName, refSeqDoc)
         }
       }
       if (!refSeqDoc) {
@@ -230,58 +215,6 @@ export abstract class FeatureChange extends Change {
         { session },
       )
       logger.verbose?.(`Added docId "${newFeatureDoc._id}"`)
-    }
-  }
-
-  /**
-   * Get children's feature ids
-   * @param feature - parent feature
-   * @returns
-   */
-  getChildFeatureIds(feature: Feature | AnnotationFeatureSnapshot): string[] {
-    if (!feature.children) {
-      return []
-    }
-    const featureIds = []
-    const children =
-      feature.children instanceof Map
-        ? feature.children
-        : new Map(Object.entries(feature.children))
-    for (const [childFeatureId, childFeature] of children || new Map()) {
-      featureIds.push(childFeatureId, ...this.getChildFeatureIds(childFeature))
-    }
-    return featureIds
-  }
-
-  /**
-   * Recursively assign new IDs to a feature
-   * @param feature - Parent feature
-   * @param featureIds -
-   */
-  generateNewIds(
-    feature: Feature | AnnotationFeatureSnapshot,
-    featureIds: string[],
-  ): AnnotationFeatureSnapshot {
-    const newId = new ObjectID().toHexString()
-    featureIds.push(newId)
-
-    const children: Record<string, AnnotationFeatureSnapshot> = {}
-    if (feature.children) {
-      Object.values(feature.children).forEach((child) => {
-        const newChild = this.generateNewIds(child, featureIds)
-        children[newChild._id] = newChild
-      })
-    }
-    const refSeq =
-      typeof feature.refSeq === 'string'
-        ? feature.refSeq
-        : (feature.refSeq as unknown as Types.ObjectId).toHexString()
-
-    return {
-      ...feature,
-      refSeq,
-      children: feature.children && children,
-      _id: newId,
     }
   }
 }
