+++ packages/apollo-shared/src/ChangeManager/changes/index.ts
@@ -1,0 +1,9 @@
+export * from './AddAssemblyAndFeaturesFromFileChange'
+export * from './AddAssemblyFromFileChange'
+export * from './AddFeatureChange'
+export * from './AddFeaturesFromFileChange'
+export * from './CopyFeatureChange'
+export * from './DeleteFeatureChange'
+export * from './LocationEndChange'
+export * from './LocationStartChange'
+export * from './TypeChange'
