+++ packages/apollo-shared/src/ChangeManager/index.ts
@@ -1,12 +1,14 @@
-import { AddAssemblyAndFeaturesFromFileChange } from './AddAssemblyAndFeaturesFromFileChange'
-import { AddAssemblyFromFileChange } from './AddAssemblyFromFileChange'
-import { AddFeatureChange } from './AddFeatureChange'
-import { AddFeaturesFromFileChange } from './AddFeaturesFromFileChange'
-import { CopyFeatureChange } from './CopyFeatureChange'
-import { DeleteFeatureChange } from './DeleteFeatureChange'
-import { LocationEndChange } from './LocationEndChange'
-import { LocationStartChange } from './LocationStartChange'
-import { TypeChange } from './TypeChange'
+import {
+  AddAssemblyAndFeaturesFromFileChange,
+  AddAssemblyFromFileChange,
+  AddFeatureChange,
+  AddFeaturesFromFileChange,
+  CopyFeatureChange,
+  DeleteFeatureChange,
+  LocationEndChange,
+  LocationStartChange,
+  TypeChange,
+} from './changes'
 
 export const changes = {
   AddAssemblyAndFeaturesFromFileChange,
@@ -20,15 +22,12 @@ export const changes = {
   TypeChange,
 }
 
-export * from './AddAssemblyAndFeaturesFromFileChange'
-export * from './AddAssemblyFromFileChange'
-export * from './AddFeaturesFromFileChange'
-export * from './Change'
+export * from './changes'
 export * from './ChangeManager'
 export * from './ChangeTypes'
-export * from './AddFeatureChange'
-export * from './CopyFeatureChange'
-export * from './DeleteFeatureChange'
-export * from './LocationEndChange'
-export * from './LocationStartChange'
-export * from './TypeChange'
+
+export {
+  SerializedChange,
+  ClientDataStore,
+  Change,
+} from './changes/abstract/Change'
