+++ packages/apollo-shared/src/Validations/Validation.ts
@@ -3,7 +3,10 @@ import type { Reflector } from '@nestjs/core'
 import { FeatureDocument, UserDocument } from 'apollo-schemas'
 import { ClientSession, Model } from 'mongoose'
 
-import { Change, ClientDataStore } from '../ChangeManager/Change'
+import {
+  Change,
+  ClientDataStore,
+} from '../ChangeManager/changes/abstract/Change'
 
 export interface Context {
   context: ExecutionContext
