+++ packages/jbrowse-plugin-apollo/src/components/ImportFeatures.tsx
@@ -107,7 +107,6 @@ export function ImportFeatures({
 
     // Add features
     const change = new AddFeaturesFromFileChange({
-      changedIds: ['1'],
       typeName: 'AddFeaturesFromFileChange',
       assemblyId,
       fileId,
